change page structure for groupe settings
