+++ webapp/pages/group/edit/_id/index.vue
@@ -1,148 +1,39 @@
 <template>
   <div>
     <ds-container>
-      <ds-form
-        class="group-form"
-        ref="groupForm"
-        v-model="formData"
-        :schema="formSchema"
-        @submit="submit"
-      >
-        <ds-input
-          v-model="formData.name"
-          label="Gruppenname"
-          placeholder="Your name ..."
-        ></ds-input>
-
-        <ds-select
-          icon="user"
-          v-model="formData.groupType"
-          label="Sichtbarkeit"
-          :options="['public', 'close', 'hidden']"
-          placeholder="Status ..."
-        ></ds-select>
-
-        <ds-input v-model="formData.about" label="Kurzbeschreibung" rows="3"></ds-input>
-
-        <ds-input
-          v-model="formData.description"
-          label="Beschreibung"
-          type="textarea"
-          rows="3"
-        ></ds-input>
-
-        <ds-select
-          icon="card"
-          v-model="formData.actionRadius"
-          label="Radius"
-          :options="['local', 'regional', 'global']"
-          placeholder="Radius ..."
-        ></ds-select>
-
-        <categories-select
-          v-if="categoriesActive"
-          model="formData.categoryIds"
-          :existingCategoryIds="formData.categoryIds"
-        />
-
-        <div>{{ formData }}</div>
-
-        <ds-space margin-top="large">
-          <ds-button @click.prevent="reset()">Reset form</ds-button>
-          <ds-button
-            type="submit"
-            @click.prevent="submit()"
-            icon="save"
-            :disabled="disabled"
-            primary
-          >
-            {{ update ? $t('group.update') : $t('group.save') }}
-          </ds-button>
-        </ds-space>
-      </ds-form>
-      <ds-space centered>
-        <nuxt-link to="/group/my-groups">zurück</nuxt-link>
-      </ds-space>
+      <group-form :group="group" :update="true" @updateGroup="updateGroup" />
     </ds-container>
   </div>
 </template>
 
 <script>
-import CategoriesSelect from '~/components/CategoriesSelect/CategoriesSelect'
+import GroupForm from '~/components/Group/GroupForm'
+import { updateGroupMutation } from '~/graphql/groups.js'
 
 export default {
-  name: 'GroupForm',
   components: {
-    CategoriesSelect,
+    GroupForm,
   },
   props: {
-    update: {
-      type: Boolean,
-      required: false,
-      default: false,
-    },
     group: {
       type: Object,
       required: false,
       default: () => ({}),
     },
   },
-  data() {
-    const { name, groupType, about, description, actionRadius, categories } = this.group
-    return {
-      categoriesActive: this.$env.CATEGORIES_ACTIVE,
-      disabled: false,
-      formData: {
-        name: name || '',
-        groupType: groupType || '',
-        about: about || '',
-        description: description || '',
-        actionRadius: actionRadius || '',
-        categoryIds: categories ? categories.map((category) => category.id) : [],
-      },
-      formSchema: {
-        name: { required: true, min: 3, max: 100 },
-        groupType: { required: true },
-        about: { required: true },
-        description: { required: true },
-        actionRadius: { required: true },
-        categoryIds: {
-          type: 'array',
-          required: this.categoriesActive,
-          validator: (_, value = []) => {
-            if (this.categoriesActive && (value.length === 0 || value.length > 3)) {
-              return [new Error(this.$t('common.validations.categories'))]
-            }
-            return []
-          },
-        },
-      },
-    }
-  },
-
   methods: {
-    submit() {
-      const { name, about, description, groupType, actionRadius, categoryIds } = this.formData
-      this.update
-        ? this.$emit('updateGroup', {
-            name,
-            about,
-            description,
-            actionRadius,
-            categoryIds,
-            id: this.group.id,
-          })
-        : this.$emit('createGroup', {
-            name,
-            about,
-            description,
-            groupType,
-            actionRadius,
-            categoryIds,
-          })
-    },
-    reset() {
-      alert('reset')
+    async updateGroup(value) {
+      const { id, name, about, description, groupType, actionRadius, categoryIds } = value
+      const variables = { id, name, about, description, groupType, actionRadius, categoryIds }
+      try {
+        await this.$apollo.mutate({
+          mutation: updateGroupMutation,
+          variables,
+        })
+        this.$toast.success(this.$t('group.group-updated'))
+      } catch (error) {
+        this.$toast.error(error.message)
+      }
     },
   },
 }
